+++ R/platemodel.R
@@ -58,6 +58,21 @@ setMethod("initialize",signature="platemodel",
 			
 				.Object@name <- ""
 			}
+			# co a comprehensive check of all files
+			# rotation
+			for( i in .Object@rotation){
+				if(!file.exists(i)) stop("The provided rotation file\n  ",i,"\n  could not be found.")
+			}
+
+			if(is.data.frame(.Object@features)){
+				checkthis <- .Object@features[,1]
+				names(checkthis) <- rownames(.Object@features)
+			}else{
+				checkthis <- .Object@features
+			}
+			for(i in checkthis){
+				if(!file.exists(i)) stop("The feature collection file\n  '",i,"'\n  could not be found.")
+			}
 
 		return(.Object)
 	}