+++ R/zzz.R
@@ -13,6 +13,9 @@
 #' @importFrom methods new
 #' @importFrom utils assignInNamespace
 #' @importFrom utils getFromNamespace
-#' @importFrom utils data 
+#' @importFrom utils data
+#' @importFrom graphics polygon
+#' @importFrom graphics lines
+#' @importFrom graphics xspline
+#' @importFrom graphics par
 NULL
-